added interactive visualization

+++ tacoma/interactive.py
@@ -16,9 +16,37 @@ import time
 import threading
 
 import wget
+import shutil
+
+import json
+from distutils.dir_util import copy_tree
 
 import tacoma as tc
 from tacoma.data_io import mkdirp_customdir
+
+standard_config = {
+    "temporal_network_files" : [
+                              ],
+    "edges_coordinate_files" : [
+                              ],
+    "plot_width" : 320 ,
+    "titles" : [
+               ],
+    "network_plot_height" : 250,
+    "edges_plot_height" : 100,
+    "padding" : 10,
+    "start_it" : 0,
+    "node_radius" : 2.5,
+    "link_distance" : 10,
+    "node_charge": -8,
+    "edge_line_width" : 1,
+    "font_size_in_px" : 14,
+    "link_width" : 1,
+    "d3_format_string": ".3f",
+
+}
+
+html_source_path = os.path.join(tc.__path__[0],'interactive')
 
 def _make_and_get_directory(path):
     directory, single = os.path.split(os.path.abspath(os.path.expanduser(path)))
@@ -29,7 +57,7 @@ def download_d3():
     """Download `d3.v4.min.js` and save it in `~/.tacoma/d3`."""
 
     url = "https://cdnjs.cloudflare.com/ajax/libs/d3/4.13.0/d3.min.js"
-    filename = "~/.tacoma/web/d3/d3.v4.min.js"
+    filename = "~/.tacoma/web/d3.v4/d3.v4.min.js"
     target_name = "d3.min.js"
     if not os.path.exists( os.path.abspath(os.path.expanduser(filename)) ):
 
@@ -43,6 +71,13 @@ def download_d3():
         # move to better name
         os.rename(os.path.join(directory, target_name), os.path.abspath(os.path.expanduser(filename)) )
 
+def prepare_visualization_directory():
+    """Move all files from the tacoma/interactive directory to ~/.tacoma/web"""
+    src = html_source_path
+    dst = os.path.abspath(os.path.expanduser("~/.tacoma/web/"))
+    if not os.path.exists(os.path.join(dst, "index.html")):
+        copy_tree(src, dst)
+
 class StoppableHTTPServer(http.server.HTTPServer):
     """ from https://stackoverflow.com/questions/268629/how-to-stop-basehttpserver-serve-forever-in-a-basehttprequesthandler-subclass """
 
@@ -53,34 +88,65 @@ class StoppableHTTPServer(http.server.HTTPServer):
         while subfolder.endswith('/'):
             subfolder = subfolder[:-1]
             
-        self.subjson = subfolder + '.json'
-
-        self.thread = threading.Thread(None, self.run)
-        self.thread.start()
+        self.subjson = subfolder + '_config.json'
 
     def run(self):
         try:
             self.serve_forever()
-        except KeyboardInterrupt:
-            self.stop_this()
-        finally:
-            self.stop_this()
+        except OSError:
+            pass
 
     def stop_this(self):
         # Clean-up server (close socket, etc.)
+        print('was asked to stop the server')
         self.server_close()
-        os.remove(self.subjson)
-        shutil.rmtree(self.subfolder)
-        self.thread.join()
 
-    def __del__(self):
-        self.stop_this()
+        #try:
+        if os.path.exists(self.subjson):
+            os.remove(self.subjson)
+        #except
+        if os.path.exists(self.subfolder):
+            try:
+                #os.rmdir(self.subfolder)
+                shutil.rmtree(self.subfolder)
+            except FileNotFoundError as e:
+                raise e
+                pass
 
-def visualize(temporal_networks,port = 8226):
+        #os.chdir(self.cwd)
+        print('deleted all files')
+
+    #def __del__(self):
+    #    self.stop_this()
+
+def _get_prepared_network(tn,dt,time_unit,time_normalization_factor):
+    
+    tn_b = tc.bin(tn,dt=dt) # rebin the network
+    tn_b.t = [ t * time_normalization_factor for t in tn_b.t ] # rescale the network's time
+    tn_b.tmax *= time_normalization_factor
+    if time_unit is None:
+        time_unit = ""
+    tn_b.time_unit = time_unit # set time unit
+
+    return tn_b
+
+def visualize(temporal_networks,
+              frame_dt,
+              time_normalization_factor = 1,
+              new_time_unit = None,
+              titles = None,
+              config = None,
+              port = 8226):
 
 
     if not hasattr(temporal_networks,'__len__'):
         temporal_networks = [ temporal_networks ]
+
+    if titles is None:
+        titles = ["" for _ in temporal_networks]
+    elif type(titles) == str or not hasattr(titles,'__len__'):
+        titles = [titles]
+    print(titles)
 
     # define the server address
     server_address = ('127.0.0.1', port)
@@ -90,38 +156,73 @@ def visualize(temporal_networks,port = 8226):
 
     # download d3 if that did not happen yet
     download_d3()
+
+    # copy the html and js files for the visualizations
+    prepare_visualization_directory()
     
     # create a subfolder based on the current time
-    subfolder = "{:x}".format(int(time.time()*1000))
+    subfolder = "tmp_{:x}".format(int(time.time()*1000))
     mkdirp_customdir(directory = web_dir)
-    mkdirp_customdir(directory = os.path.join(web_dir,subfolder))
+    subfolder_path = os.path.join(web_dir,subfolder)
+    mkdirp_customdir(directory = subfolder_path)
 
     # change directory to this directory
-    print("starting server at", web_dir)
+    print("changing directory to", web_dir)
+    print("starting server here ...", web_dir)
     cwd = os.getcwd()
     os.chdir(web_dir)
 
     server = StoppableHTTPServer(("127.0.0.1", 8226),
                                   http.server.SimpleHTTPRequestHandler,
-                                  os.path.join(web_dir,subfolder)
+                                  subfolder_path,
                                   )
-    # thread = threading.Thread(None, server.run)
-    # thread.start()
 
-    webbrowser.open("http://localhost:8226/")
+    for itn, tn in enumerate(temporal_networks):
+        print("preparing network", titles[itn])
+        tn_b = _get_prepared_network(tn, frame_dt, new_time_unit, time_normalization_factor)
+        taco_fname = os.path.join(subfolder, subfolder+'_'+str(itn)+'.taco')
+        edge_fname = os.path.join(subfolder, subfolder+'_'+str(itn)+'.json')
+        tc.write_edge_trajectory_coordinates(tn_b,
+                                     os.path.join(web_dir, edge_fname),
+                                     filter_for_duration = frame_dt * time_normalization_factor)
+        tc.write_json_taco(tn_b, os.path.join(web_dir,taco_fname))
+        
+        standard_config['temporal_network_files'].append( taco_fname )
+        standard_config['edges_coordinate_files'].append( edge_fname )
+        standard_config['titles'].append( titles[itn] )
+
+
+    if type(config) == dict:
+        standard_config.update(config)
+
+    with open(os.path.join(web_dir, subfolder+'_config.json' ),'w') as f:
+        json.dump(standard_config, f)
+
+
+    # ========= start server ============
+    thread = threading.Thread(None, server.run)
+    thread.start()
+
+    webbrowser.open("http://localhost:8226/?data=" + subfolder)
 
     try:
         while True:
             time.sleep(2)
     except KeyboardInterrupt:
+        #thread.join()
+        print('stopping server ...')
         server.stop_this()
+        #thread.join()
 
     #time.sleep(1)
-    #os.chdir(cwd)
+
+    print('changing directory back to', cwd)
+
+    os.chdir(cwd)
 
 if __name__ == "__main__":
     #download_d3()
     a = tc.load_json_taco("~/.tacoma/ht09.taco")
-    visualize(a)
+    visualize(a,frame_dt = 20,titles = 'HT09',new_time_unit='h',time_normalization_factor=1./3600.)
